connectors/bright_data_unlocker/helpers/schema_management.py
"""Schema and fields management utilities."""

from pathlib import Path
from typing import Set

import yaml

from fivetran_connector_sdk import Logging as log

FIELDS_FILE = Path("fields.yaml")


def _load_existing_tables() -> dict:
    if not FIELDS_FILE.exists():
        return {"tables": {}}

    try:
        with FIELDS_FILE.open("r", encoding="utf-8") as yaml_file:
            data = yaml.safe_load(yaml_file) or {}
    except (IOError, yaml.YAMLError) as exc:
        log.info(f"Warning: Could not read existing fields.yaml: {str(exc)}")
        return {"tables": {}}

    tables = data.get("tables", {})
    if not isinstance(tables, dict):
        log.info("Warning: Unexpected fields.yaml structure. Resetting tables section.")
        tables = {}
    return {"tables": tables}


def _default_description(table_name: str) -> str:
    return (
        f"Dynamically created table '{table_name}' from Bright Data results. "
        "Fields are inferred from the API response structure."
    )


def update_fields_yaml(fields: Set[str], table_name: str) -> None:
    """
    Update fields.yaml with discovered fields from Bright Data results.

    Documents all fields found in the API responses for reference.

    Args:
        fields: Set of field names discovered from results
        table_name: Name of the table these fields belong to
    """
    sorted_fields = sorted(list(fields))

    fields_data = _load_existing_tables()
    tables = fields_data.setdefault("tables", {})
    table_entry = tables.setdefault(
        table_name,
        {
            "description": _default_description(table_name),
            "fields": [],
        },
    )

    table_entry["description"] = table_entry.get(
        "description", _default_description(table_name)
    )
    table_entry["fields"] = sorted_fields

    try:
        with FIELDS_FILE.open("w", encoding="utf-8") as yaml_file:
            yaml.safe_dump(
                fields_data,
                yaml_file,
                default_flow_style=False,
                sort_keys=True,
                allow_unicode=True,
            )
        log.info(
            f"Updated fields.yaml with {len(fields)} fields for table '{table_name}'"
        )
    except (IOError, yaml.YAMLError) as exc:
        log.info(f"Warning: Could not update fields.yaml: {str(exc)}")
